requirements.txt
# ==== Dash App Requirments ==== 
dash==3.2.0
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3